modules: add failsafe module
The failsafe module adds a minimal web-interface which is intended to
run on a very compact initramfs-based image intended only for flashing
firmware onto a device. This can be used as part of an installation method
or reside in a small partition on the device flash which can be booted
in case the checksum-test of the main kernel or rootfs fails.

Signed-off-by: Daniel Golle <dgolle@allnet.de>

git-svn-id: http://svn.luci.subsignal.org/luci@9310 ab181a69-ba2e-0410-a84d-ff88ab4c47bc