[luci] Fixes the bug that  some table will delete the header row in IE
It will insert some rows according to the responding content.
The IE will insert them into the tbody, but the header row is not
in the tbody, the last row become the rows before the header row.
add tbody tag in the table tag, make the header row inside the tbody.

CRs-fixed:461452

Change-Id: Ia1674f2ff6d02676c6f6753d4edd95f59e7f9420
Signed-off-by: ken zhu <guigenz@codeaurora.org>