luci: fix crash when add new wifi vif interface
The web interface will crash when add a new wifi vif interface.
It is because it will invoke a non exist method `is_up` of a lua class entity `found_sta`.

It was introduced by this commit which merged some buggy patches:
c9d9415 (luci: Merge wifi related luci patches, 2016-07-08)

The `is_up` and `channel` should be members of `found_sta` not a method
of it. So modify the patch to fix this.

CRs-Fixed: 3834565
Change-Id: I2df730ea273366d625d2413b1a9197470ddfc63d
Signed-off-by: Long Cheng <quic_longchen@quicinc.com>